Linux: Hoist identical code from GTK and Portal implementations into Linux shared header
